+++ src/core/camera.rs
@@ -3,8 +3,8 @@ use core::quat;
 use core::ray;
 use core::vector;
 
-pub const HORIZONTAL_APERTURE_35MM: f64 = 2.2;
-pub const VERTICAL_APERTURE_35MM: f64 = 1.6;
+pub const HORIZONTAL_APERTURE_35MM: f32 = 2.2;
+pub const VERTICAL_APERTURE_35MM: f32 = 1.6;
 
 /** Perspective camera representation. */
 pub struct Camera {
@@ -12,22 +12,22 @@ pub struct Camera {
      * The distance from the eye to the focal plane.
      * A longer focal length means greater magnification, and vice versa.
      */
-    pub focal_length: f64,
+    pub focal_length: f32,
     /**
      * The width of the projector aperture.
      * This must be in the same units as other scene dimensions, e.g. for a 35mm camera in a
      * cm-world, you should use 3.5cm (but really 3.6cm because it's actually 36mm...).
      */
-    pub horizontal_aperture: f64,
+    pub horizontal_aperture: f32,
     /**
      * The height of the projector aperture. See the documentation for horizontal_aperture
      * for the same unit restrictions. */
-    pub vertical_aperture: f64,
+    pub vertical_aperture: f32,
     /**
      * The f-number or focal ratio. A larger f-stop gives more depth of field, bringing more
      * objects into focus. A smaller f-stop will narrow the focus around the focal length.
      */
-    pub f_stop: f64,
+    pub f_stop: f32,
     pub xform: matrix::Mat,
 }
 
@@ -38,10 +38,10 @@ impl Camera {
     }
 
     pub fn new(
-        focal_length: f64,
-        horizontal_aperture: f64,
-        vertical_aperture: f64,
-        f_stop: f64,
+        focal_length: f32,
+        horizontal_aperture: f32,
+        vertical_aperture: f32,
+        f_stop: f32,
         rotate: &quat::Quat,
         translate: &vector::Vec) -> Camera
     {
@@ -62,16 +62,16 @@ impl Camera {
      * The radius of the entrance pupil.
      * See <https://en.wikipedia.org/wiki/F-number> for derivation.
      */
-    pub fn pupil_radius(&self) -> f64 {
+    pub fn pupil_radius(&self) -> f32 {
         0.5 * (self.focal_length / self.f_stop)
     }
 
     /** Horizontal to vertical ratio. */
-    pub fn aspect_ratio(&self) -> f64 {
+    pub fn aspect_ratio(&self) -> f32 {
         self.horizontal_aperture / self.vertical_aperture
     }
 
-    pub fn window_max(&self) -> (f64, f64) {
+    pub fn window_max(&self) -> (f32, f32) {
         (self.horizontal_aperture / (self.focal_length * 2.0),
          self.vertical_aperture / (self.focal_length * 2.0))
     }
@@ -82,7 +82,7 @@ impl Camera {
      * center, (-1, 1) is the lower-left, and (1, 1) is the upper-right.
      * Other documentation may refer to these types of coordinates as being in "lens space".
      */
-    pub fn compute_ray(&self, s: f64, t: f64) -> ray::Ray {
+    pub fn compute_ray(&self, s: f32, t: f32) -> ray::Ray {
         let window_max = self.window_max();
         let origin = vector::Vec::zero();
         let direction = vector::Vec::new(window_max.0 * s, window_max.1 * t, -1.0)
